Debugging, test calss
Debugging for login, test classes
