+++ src/main/webapp/AddDonation.jsp
@@ -16,6 +16,7 @@
 	 if (successMessage.trim() !== "") {
 	     // Display the success message as a pop-up
 	     alert(successMessage);
+	 }
  </script>
 </head>
 <body>
