+++ src/main/webapp/AdminDonorProfile.jsp
@@ -32,6 +32,8 @@
     if (successMessage.trim() !== "") {
         // Display the success message as a pop-up
         alert(successMessage);
+        }
+    
     </script>
 </head>
 <body>
