+++ src/main/webapp/DonorAllCampaigns.jsp
@@ -27,6 +27,8 @@
     if (successMessage.trim() !== "") {
         // Display the success message as a pop-up
         alert(successMessage);
+        }
+    
     </script>
 </head>
 <body>
