+++ src/main/webapp/DonorProfile.jsp
@@ -35,6 +35,7 @@
     if (successMessage.trim() !== "") {
         // Display the success message as a pop-up
         alert(successMessage);
+    }
     </script>
 </head>
 <body>
