+++ src/main/webapp/UpdateDonorDetails.jsp
@@ -75,6 +75,7 @@
         if (successMessage.trim() !== "") {
             // Display the success message as a pop-up
             alert(successMessage);
+        }
     </script>
 </head>
 <body>
