+++ src/main/webapp/campaignUpdateDelete.jsp
@@ -26,6 +26,8 @@
     if (successMessage.trim() !== "") {
         // Display the success message as a pop-up
         alert(successMessage);
+        }
+    
     
 </script>
 <style>
