+++ src/main/webapp/forgetPassword.jsp
@@ -64,6 +64,7 @@
         if (successMessage.trim() !== "") {
             // Display the success message as a pop-up
             alert(successMessage);
+        }
         
     </script>
 </head>
