+++ src/main/webapp/registration.jsp
@@ -32,6 +32,7 @@
         if (successMessage.trim() !== "") {
             // Display the success message as a pop-up
             alert(successMessage);
+        }
     </script>
 </head>
 <body>
