+++ src/main/webapp/updateCampaign.jsp
@@ -17,6 +17,7 @@
     if (successMessage.trim() !== "") {
         // Display the success message as a pop-up
         alert(successMessage);
+    }
         </script>
 </head>
 <body>
