+++ test-output/Default suite/Default test.html
@@ -55,11 +55,11 @@ function toggleAllBoxes() {
 <body>
 <h2 align='center'>Default test</h2><table border='1' align="center">
 <tr>
-<td>Tests passed/Failed/Skipped:</td><td>2/0/0</td>
+<td>Tests passed/Failed/Skipped:</td><td>7/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Mon Apr 29 22:42:05 GST 2024</td>
+<td>Started on:</td><td>Tue Apr 30 22:12:25 GST 2024</td>
 </tr>
-<tr><td>Total time:</td><td>0 seconds (268 ms)</td>
+<tr><td>Total time:</td><td>0 seconds (588 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -76,15 +76,40 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='Test.DonorCRUDTest.testGetDonor()'><b>testGetDonor</b><br>Test class: Test.DonorCRUDTest</td>
+<td title='Test.UnitTest.testgetAdminType()'><b>testgetAdminType</b><br>Test class: Test.UnitTest</td>
 <td></td>
 <td>0</td>
-<td>Test.DonorCRUDTest@3a7442c7</td></tr>
+<td>Test.UnitTest@cb0755b</td></tr>
 <tr>
-<td title='Test.DonorCRUDTest.testNicUsed()'><b>testNicUsed</b><br>Test class: Test.DonorCRUDTest</td>
+<td title='Test.UnitTest.testDonationDeleted()'><b>testDonationDeleted</b><br>Test class: Test.UnitTest</td>
 <td></td>
 <td>0</td>
-<td>Test.DonorCRUDTest@3a7442c7</td></tr>
+<td>Test.UnitTest@cb0755b</td></tr>
+<tr>
+<td title='Test.UnitTest.testgetDonorType()'><b>testgetDonorType</b><br>Test class: Test.UnitTest</td>
+<td></td>
+<td>0</td>
+<td>Test.UnitTest@cb0755b</td></tr>
+<tr>
+<td title='Test.UnitTest.testNicUsed()'><b>testNicUsed</b><br>Test class: Test.UnitTest</td>
+<td></td>
+<td>0</td>
+<td>Test.UnitTest@cb0755b</td></tr>
+<tr>
+<td title='Test.UnitTest.testCampaignDeleted()'><b>testCampaignDeleted</b><br>Test class: Test.UnitTest</td>
+<td></td>
+<td>0</td>
+<td>Test.UnitTest@cb0755b</td></tr>
+<tr>
+<td title='Test.UnitTest.testLogin()'><b>testLogin</b><br>Test class: Test.UnitTest</td>
+<td></td>
+<td>0</td>
+<td>Test.UnitTest@cb0755b</td></tr>
+<tr>
+<td title='Test.UnitTest.testGetDonor()'><b>testGetDonor</b><br>Test class: Test.UnitTest</td>
+<td></td>
+<td>0</td>
+<td>Test.UnitTest@cb0755b</td></tr>
 </table><p>
 </body>
 </html>
