+++ test-output/Default suite/Default test.xml
@@ -1,6 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="0" hostname="Dimal_ROG" failures="0" tests="2" name="Default test" time="0.268" errors="0" timestamp="2024-04-29T22:42:06 GST">
-  <testcase classname="Test.DonorCRUDTest" name="testGetDonor" time="0.208"/>
-  <testcase classname="Test.DonorCRUDTest" name="testNicUsed" time="0.007"/>
+<testsuite ignored="0" hostname="Dimal_ROG" failures="0" tests="7" name="Default test" time="0.588" errors="0" timestamp="2024-04-30T22:12:25 GST">
+  <testcase classname="Test.UnitTest" name="testCampaignDeleted" time="0.173"/>
+  <testcase classname="Test.UnitTest" name="testDonationDeleted" time="0.007"/>
+  <testcase classname="Test.UnitTest" name="testGetDonor" time="0.03"/>
+  <testcase classname="Test.UnitTest" name="testLogin" time="0.303"/>
+  <testcase classname="Test.UnitTest" name="testNicUsed" time="0.005"/>
+  <testcase classname="Test.UnitTest" name="testgetAdminType" time="0.007"/>
+  <testcase classname="Test.UnitTest" name="testgetDonorType" time="0.007"/>
 </testsuite> <!-- Default test -->
