+++ test-output/emailable-report.html
@@ -9,11 +9,16 @@
 <table>
 <tr><th>Test</th><th># Passed</th><th># Skipped</th><th># Retried</th><th># Failed</th><th>Time (ms)</th><th>Included Groups</th><th>Excluded Groups</th></tr>
 <tr><th colspan="8">Default suite</th></tr>
-<tr><td><a href="#t0">Default test</a></td><td class="num">2</td><td class="num">0</td><td class="num">0</td><td class="num">0</td><td class="num">268</td><td></td><td></td></tr>
+<tr><td><a href="#t0">Default test</a></td><td class="num">7</td><td class="num">0</td><td class="num">0</td><td class="num">0</td><td class="num">588</td><td></td><td></td></tr>
 </table>
-<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Default suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Default test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="2">Test.DonorCRUDTest</td><td><a href="#m0">testGetDonor</a></td><td rowspan="1">1714416125848</td><td rowspan="1">208</td></tr><tr class="passedeven"><td><a href="#m1">testNicUsed</a></td><td rowspan="1">1714416126060</td><td rowspan="1">7</td></tr></tbody>
+<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Default suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Default test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="7">Test.UnitTest</td><td><a href="#m0">testCampaignDeleted</a></td><td rowspan="1">1714500745057</td><td rowspan="1">173</td></tr><tr class="passedeven"><td><a href="#m1">testDonationDeleted</a></td><td rowspan="1">1714500745233</td><td rowspan="1">7</td></tr><tr class="passedeven"><td><a href="#m2">testGetDonor</a></td><td rowspan="1">1714500745241</td><td rowspan="1">30</td></tr><tr class="passedeven"><td><a href="#m3">testLogin</a></td><td rowspan="1">1714500745273</td><td rowspan="1">303</td></tr><tr class="passedeven"><td><a href="#m4">testNicUsed</a></td><td rowspan="1">1714500745578</td><td rowspan="1">5</td></tr><tr class="passedeven"><td><a href="#m5">testgetAdminType</a></td><td rowspan="1">1714500745584</td><td rowspan="1">7</td></tr><tr class="passedeven"><td><a href="#m6">testgetDonorType</a></td><td rowspan="1">1714500745592</td><td rowspan="1">7</td></tr></tbody>
 </table>
-<h2>Default test</h2><h3 id="m0">Test.DonorCRUDTest#testGetDonor</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
-<h3 id="m1">Test.DonorCRUDTest#testNicUsed</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h2>Default test</h2><h3 id="m0">Test.UnitTest#testCampaignDeleted</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m1">Test.UnitTest#testDonationDeleted</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m2">Test.UnitTest#testGetDonor</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m3">Test.UnitTest#testLogin</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m4">Test.UnitTest#testNicUsed</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m5">Test.UnitTest#testgetAdminType</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m6">Test.UnitTest#testgetDonorType</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
 </body>
 </html>
