+++ test-output/junitreports/TEST-Test.UnitTest.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="Dimal_ROG" failures="0" tests="7" name="Test.UnitTest" time="0.532" errors="0" timestamp="2024-04-30T22:12:25 GST" skipped="0">
+  <testcase classname="Test.UnitTest" name="testgetAdminType" time="0.007"/>
+  <testcase classname="Test.UnitTest" name="testDonationDeleted" time="0.007"/>
+  <testcase classname="Test.UnitTest" name="testgetDonorType" time="0.007"/>
+  <testcase classname="Test.UnitTest" name="testNicUsed" time="0.005"/>
+  <testcase classname="Test.UnitTest" name="testCampaignDeleted" time="0.173"/>
+  <testcase classname="Test.UnitTest" name="testLogin" time="0.303"/>
+  <testcase classname="Test.UnitTest" name="testGetDonor" time="0.030"/>
+</testsuite> <!-- Test.UnitTest -->
