+++ test-output/testng-results.xml
@@ -1,21 +1,41 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<testng-results ignored="0" total="2" passed="2" failed="0" skipped="0">
+<testng-results ignored="0" total="7" passed="7" failed="0" skipped="0">
   <reporter-output>
   </reporter-output>
-  <suite started-at="2024-04-29T22:42:05 GST" name="Default suite" finished-at="2024-04-29T22:42:06 GST" duration-ms="268">
+  <suite started-at="2024-04-30T22:12:25 GST" name="Default suite" finished-at="2024-04-30T22:12:25 GST" duration-ms="588">
     <groups>
     </groups>
-    <test started-at="2024-04-29T22:42:05 GST" name="Default test" finished-at="2024-04-29T22:42:06 GST" duration-ms="268">
-      <class name="Test.DonorCRUDTest">
-        <test-method signature="testGetDonor()[pri:0, instance:Test.DonorCRUDTest@3a7442c7]" started-at="2024-04-29T22:42:05 GST" name="testGetDonor" finished-at="2024-04-29T22:42:06 GST" duration-ms="208" status="PASS">
+    <test started-at="2024-04-30T22:12:25 GST" name="Default test" finished-at="2024-04-30T22:12:25 GST" duration-ms="588">
+      <class name="Test.UnitTest">
+        <test-method signature="testCampaignDeleted()[pri:0, instance:Test.UnitTest@cb0755b]" started-at="2024-04-30T22:12:25 GST" name="testCampaignDeleted" finished-at="2024-04-30T22:12:25 GST" duration-ms="173" status="PASS">
+          <reporter-output>
+          </reporter-output>
+        </test-method> <!-- testCampaignDeleted -->
+        <test-method signature="testDonationDeleted()[pri:0, instance:Test.UnitTest@cb0755b]" started-at="2024-04-30T22:12:25 GST" name="testDonationDeleted" finished-at="2024-04-30T22:12:25 GST" duration-ms="7" status="PASS">
+          <reporter-output>
+          </reporter-output>
+        </test-method> <!-- testDonationDeleted -->
+        <test-method signature="testGetDonor()[pri:0, instance:Test.UnitTest@cb0755b]" started-at="2024-04-30T22:12:25 GST" name="testGetDonor" finished-at="2024-04-30T22:12:25 GST" duration-ms="30" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- testGetDonor -->
-        <test-method signature="testNicUsed()[pri:0, instance:Test.DonorCRUDTest@3a7442c7]" started-at="2024-04-29T22:42:06 GST" name="testNicUsed" finished-at="2024-04-29T22:42:06 GST" duration-ms="7" status="PASS">
+        <test-method signature="testLogin()[pri:0, instance:Test.UnitTest@cb0755b]" started-at="2024-04-30T22:12:25 GST" name="testLogin" finished-at="2024-04-30T22:12:25 GST" duration-ms="303" status="PASS">
+          <reporter-output>
+          </reporter-output>
+        </test-method> <!-- testLogin -->
+        <test-method signature="testNicUsed()[pri:0, instance:Test.UnitTest@cb0755b]" started-at="2024-04-30T22:12:25 GST" name="testNicUsed" finished-at="2024-04-30T22:12:25 GST" duration-ms="5" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- testNicUsed -->
-      </class> <!-- Test.DonorCRUDTest -->
+        <test-method signature="testgetAdminType()[pri:0, instance:Test.UnitTest@cb0755b]" started-at="2024-04-30T22:12:25 GST" name="testgetAdminType" finished-at="2024-04-30T22:12:25 GST" duration-ms="7" status="PASS">
+          <reporter-output>
+          </reporter-output>
+        </test-method> <!-- testgetAdminType -->
+        <test-method signature="testgetDonorType()[pri:0, instance:Test.UnitTest@cb0755b]" started-at="2024-04-30T22:12:25 GST" name="testgetDonorType" finished-at="2024-04-30T22:12:25 GST" duration-ms="7" status="PASS">
+          <reporter-output>
+          </reporter-output>
+        </test-method> <!-- testgetDonorType -->
+      </class> <!-- Test.UnitTest -->
     </test> <!-- Default test -->
   </suite> <!-- Default suite -->
 </testng-results>
